Small fixes in 17 Observables
