+++ 17 Observables/src/app.tsx
@@ -1,9 +1,8 @@
-import * as React from "react";
-import { HelloWorldContainer } from "./components/helloworld";
-import { NameEditContainer } from "./components/nameEdit";
-import { ColorDisplayerContainer } from "./components/color";
-import { ColorPickerContainer } from "./components/color";
-import {MembersAreaContainer} from "./components/members";
+import * as React from 'react';
+import { HelloWorldContainer } from './components/helloworld'
+import { NameEditContainer } from './components/nameEdit';
+import { ColorDisplayerContainer, ColorPickerContainer } from './components/color';
+import { MembersAreaContainer } from './components/members';
 
 export const App = () => {
   return (
@@ -19,4 +18,4 @@ export const App = () => {
       <ColorPickerContainer/>
     </div>
   );
-};
+}
