+++ 17 Observables/src/components/members/index.ts
@@ -1,4 +1,4 @@
-import {MembersAreaContainer} from './memberAreaContainer';
+import { MembersAreaContainer } from './memberAreaContainer';
 
 export {
   MembersAreaContainer
