+++ 17 Observables/src/components/members/memberAreaContainer.ts
@@ -1,21 +1,20 @@
 import { connect } from 'react-redux';
 import { memberRequest } from '../../actions/memberRequest';
-import { MembersArea } from './memberarea';
-
+import { MembersArea } from './memberArea';
 
 const mapStateToProps = (state) => {
-    return{
-        members: state.memberReducer.members
-    };
+  return{
+    members: state.memberReducer.members
+  };
 }
 
 const mapDispatchToProps = (dispatch) => {
-    return {
-        loadMembers: () => {return dispatch(memberRequest())}
-    };
+  return {
+    loadMembers: () => {return dispatch(memberRequest())}
+  };
 }
 
 export const MembersAreaContainer = connect(
-    mapStateToProps,
-    mapDispatchToProps
+  mapStateToProps,
+  mapDispatchToProps
 )(MembersArea)
