--- 17 Observables/src/components/members/memberrow.tsx
@@ -1,23 +1,0 @@
-import * as React from 'react';
-import {MemberEntity} from '../../model/member';
-
-
-interface Props  {
-  member : MemberEntity;
-}
-
-export const MemberRow = (props: Props) => {
-     return (
-       <tr>
-         <td>
-           <img src={props.member.avatar_url} className="avatar"/>
-         </td>
-         <td>
-           <span>{props.member.id}</span>
-         </td>
-         <td>
-           <span>{props.member.login}</span>
-         </td>
-       </tr>
-     );
-}
