+++ 17 Observables/src/reducers/index.ts
@@ -1,6 +1,6 @@
-import { combineReducers } from "redux";
-import { userProfileReducer } from "./userProfile";
-import { memberReducer } from "./memberReducer";
+import { combineReducers } from 'redux';
+import { userProfileReducer } from './userProfile';
+import { memberReducer } from './memberReducer';
 
 export const reducers =  combineReducers({
   userProfileReducer,
