+++ 17 Observables/src/reducers/memberReducer.ts
@@ -1,27 +1,26 @@
-import {actionsEnums} from '../common/actionsEnums';
-import {MemberEntity} from '../model/member';
-import objectAssign = require('object-assign');
+import { actionsEnums } from "../common/actionsEnums";
+import { MemberEntity } from "../model/member";
+import objectAssign = require("object-assign");
 
 class memberState  {
-  members : MemberEntity[];
+  members: MemberEntity[];
 
-  public constructor()
-  {
+  public constructor() {
     this.members = [];
   }
 }
 
-export const memberReducer =  (state : memberState = new memberState(), action) => {
-      switch (action.type) {
-        case actionsEnums.MEMBER_REQUEST_COMPLETED:
-           return handleMemberRequestCompletedAction(state, action);
-      }
+export const memberReducer =  (state: memberState = new memberState(), action) => {
+  switch (action.type) {
+    case actionsEnums.MEMBER_REQUEST_COMPLETED:
+      return handleMemberRequestCompletedAction(state, action);
+  }
 
-      return state;
+  return state;
 };
 
 
-const handleMemberRequestCompletedAction = (state : memberState, action) => {
+const handleMemberRequestCompletedAction = (state: memberState, action) => {
   const newState = objectAssign({}, state, {members: action.members});
   return newState;
 }
